Fixed error and added a component  #309
Now if the table has dependencies it display an error message. Fixed tests as well, more tests need to be added in the future
